feat: implement HoverPopover component for reusable hover-triggered popovers
